+++ Academy/Services/LecturerService.cs
@@ -10,41 +10,82 @@ namespace Services
     {
         private readonly AcademyRepository academyRepository;
 
+        public LecturerService()
+        {
+        }
+
         public LecturerService(AcademyRepository academyRepository)
         {
             this.academyRepository = academyRepository;
         }
 
-        public List<Lecturer> GetAllLecturers()
+        public virtual List<Lecturer> GetAllLecturers()
         {
+            if (academyRepository == null)
+            {
+                return new List<Lecturer>();
+            }
             List<Lecturer> lecturersList = academyRepository.GetAllLecturers();
             return lecturersList;
         }
 
-        public Lecturer GetLecturer(int id)
+        public virtual Lecturer GetLecturer(int id)
         {
+            if (academyRepository == null)
+            {
+                return null;
+            }
             Lecturer lecturer = academyRepository.GetLecturer(id);
             return lecturer;
         }
 
-        public void CreateLecturer(Lecturer lecturer)
+        public virtual bool CreateLecturer(Lecturer lecturer)
         {
+            if (academyRepository == null)
+            {
+                return false;
+            }
             academyRepository.CreateLecturer(lecturer);
+            return true;
         }
 
-        public void UpdateLecturer(Lecturer lecturer)
+        public virtual bool UpdateLecturer(Lecturer lecturer)
         {
+            if (academyRepository == null)
+            {
+                return false;
+            }
             academyRepository.UpdateLecturer(lecturer);
+            return true;
         }
 
-        public void DeleteLecturer(Lecturer lecturer)
+        public virtual bool DeleteLecturer(Lecturer lecturer)
         {
+            if (academyRepository == null)
+            {
+                return false;
+            }
             academyRepository.DeleteLecturer(lecturer);
+            return true;
         }
 
-        public void DeleteLecturer(int id)
+        public virtual bool DeleteLecturer(int id)
         {
+            if (academyRepository == null)
+            {
+                return false;
+            }
             academyRepository.DeleteLecturer(id);
+            return true;
+        }
+
+        public virtual bool IsLecturerExisted(Lecturer lecturer)
+        {
+            Lecturer dBlecturer = GetLecturer(lecturer.Id);
+            if (dBlecturer == null)
+                return false;
+
+            return dBlecturer.Equals(lecturer);
         }
     }
 }
